add message support

+++ python/lwqq/lwqq.py
@@ -43,7 +43,7 @@ class Lwqq(object):
                 ('msg_list',c_voidp),
 
                 ('msg_id',c_long),
-                ('stat',c_long),
+                ('stat',c_int),
 
                 ('dispatch',DISPATCH_FUNC),
                 ]
